+++ src/sessions/screen-share-session-handler.ts
@@ -5,8 +5,6 @@ import { startDisplayMedia, checkAllTracksHaveEnded } from '../media-utils';
 import { throwSdkError, parseJwt, isAcdJid } from '../utils';
 
 export default class ScreenShareSessionHandler extends BaseSessionHandler {
-  temporaryOutboundStream: MediaStream;
-
   sessionType = SessionTypes.acdScreenShare;
 
   shouldHandleSessionByJid (jid: string): boolean {
@@ -17,10 +15,8 @@ export default class ScreenShareSessionHandler extends BaseSessionHandler {
   async startSession (startParams: IStartSessionParams): Promise<{ conversationId: string }> {
     const { jwt, conversation, sourceCommunicationId } = this.sdk._customerData;
 
-    const stream = await startDisplayMedia();
     const jid = parseJwt(jwt).data.jid;
     const opts = {
-      stream,
       jid,
       conversationId: conversation.id,
       sourceCommunicationId: sourceCommunicationId,
@@ -30,7 +26,6 @@ export default class ScreenShareSessionHandler extends BaseSessionHandler {
     this.log(LogLevels.info, 'starting acd screen share session', opts);
 
     this.sdk._streamingConnection.webrtcSessions.initiateRtcSession(opts);
-    this.temporaryOutboundStream = stream;
     return { conversationId: conversation.id };
   }
 
@@ -47,39 +42,41 @@ export default class ScreenShareSessionHandler extends BaseSessionHandler {
   }
 
   async handleSessionInit (session: IJingleSession): Promise<void> {
-    await super.handleSessionInit(session);
+    try {
+      await super.handleSessionInit(session);
+    } catch (err) {
+      this.log(LogLevels.error, 'Screen share session init failed');
+      await super.endSession(session);
+    } finally {
+      const stream = await startDisplayMedia();
 
-    session.on('terminated', (session: IJingleSession) => {
-      session._screenShareStream.getTracks().forEach((t: MediaStreamTrack) => t.stop());
-    });
+      session.on('terminated', (session: IJingleSession) => {
+        session._screenShareStream.getTracks().forEach((t: MediaStreamTrack) => t.stop());
+      });
 
-    if (!this.sdk.isGuest) {
-      throwSdkError.call(this.sdk, SdkErrorTypes.not_supported, 'Screen share sessions not supported for authenticated users');
-    }
+      if (!this.sdk.isGuest) {
+        throwSdkError.call(this.sdk, SdkErrorTypes.not_supported, 'Screen share sessions not supported for authenticated users');
+      }
 
-    if (this.temporaryOutboundStream) {
-      this.temporaryOutboundStream.getTracks().forEach((track: MediaStreamTrack) => {
+      stream.getTracks().forEach((track: MediaStreamTrack) => {
         track.addEventListener('ended', this.onTrackEnd.bind(this, session));
       });
-      this.log(LogLevels.debug, 'temporaryOutboundStream exists. Adding stream to the session and setting it to _screenShareStream');
+      this.log(LogLevels.debug, 'Adding stream to the session and setting it to _screenShareStream', { sessionId: session.id, conversationId: session.conversationId });
 
-      await this.addMediaToSession(session, this.temporaryOutboundStream);
+      await this.addMediaToSession(session, stream);
 
-      session._screenShareStream = this.temporaryOutboundStream;
-      this.temporaryOutboundStream = null;
-    } else {
-      this.log(LogLevels.warn, 'There is no `temporaryOutboundStream` for guest user');
+      session._screenShareStream = stream;
+
+      if (!this.sdk._config.autoConnectSessions) {
+        // if autoConnectSessions is 'false' and we have a guest, throw an error
+        //  guests should auto accept screen share session
+        const errMsg = '`autoConnectSession` must be set to "true" for guests';
+        this.log(LogLevels.error, errMsg, { sessionId: session.id, conversationId: session.conversationId });
+        throwSdkError.call(this.sdk, SdkErrorTypes.generic, errMsg);
+      }
+
+      await this.acceptSession(session, { id: session.id });
     }
-
-    if (!this.sdk._config.autoConnectSessions) {
-      // if autoConnectSessions is 'false' and we have a guest, throw an error
-      //  guests should auto accept screen share session
-      const errMsg = '`autoConnectSession` must be set to "true" for guests';
-      this.log(LogLevels.error, errMsg);
-      throwSdkError.call(this.sdk, SdkErrorTypes.generic, errMsg);
-    }
-
-    await this.acceptSession(session, { id: session.id });
   }
 
   public updateOutgoingMedia (session: IJingleSession, options: IUpdateOutgoingMedia): never {